Get rid of the Bytes new-type
While it's a good idea, there is no need to have it, because
using Vec<u8> is more natural, and Serialization was not even
using it (when it should if it exists).

If Telegram ever adds a byte type which occuppies one byte
and deserialization is to be implemented, then Vec<T> where
T is u8 should behave differently to the now Vec<u8>. This
is impossible, which is why the new-type is a good idea,
because byte-strings are different from generic vectors
containing bytes.