Improve error situation (#6)
Previously, a lot of errors were just converted into an IO error
without much thought to get things going. This makes it hard to
have strongly typed errors and make use of them.

Now, errors come in their own struct and enum, and although it's
a bit more painful to work with, it's a lot less careless and a
good step forward.