Change type serialization to work on Vec<u8>, not Write
The problem with std::io::Write is that it may fail with
std::io::Error (possibly not found, permission denied,
timed out, etc.) when in practice none of these errors
happen.

By forcing to write to an in-memory buffer (just a Vec)
we can guarantee serialization will never fail, which
is a nice property.