Calculate the first salt from authentication.
`new_nonce ^ server_nonce` is used as the first salt. Calculating this
locally means the first message sent after a new connection won't be
rejected for having an incorrect salt and have to be retried.

Instead of making `create_key` return a 3-tuple to include the new
data, there's a new struct `Finished` to bundle all the data together.