Split Session into trait, memory and file impl
The idea is that users will be responsible for saving the session
file however they want to, whenever they want to. However, we want
to provide a sensible default with a file version.

Because the memory and file sessions are different (one loads and
saves to memory, the other to a file), it's better to put them as
different types and unify them with a trait.