Fix MessageBox not restoring gaps when failing to reapply
If process_updates had things in possible_gap and apply_pts_info
failed, the update would not be re-inserted in possible_gap because
there was already a gap deadline.

The right approach here is to set a new deadline if there isn't one
already, but don't change the existing one if it already exists,
because it should finish as early as the earliest update needs.

However, the update should be inserted into the gap always. If it's
not re-inserted, then process_updates would think that the gap was
resolved.

This likely happens when the gap is more than 1 so that applying
the update fails.