Always return a chat from Message::Chat
At the very least, the Chat will *always* contain the ID.
The access hash should be contained by the cache (which is filled when
missing via updates.getDifference).

If all of that fails however, still return something, even if unusable.
Message::chat() is the only way to access the chat, and it should always
provide the Chat::id().