Fix iter_participants offset not being updated
Closes #124.
The check to avoid updating the offset should be done *after*
the buffer is filled (or else it will always be empty), but it
probably was not done because the participants value is moved.

Saving the values for the check before filling the buffer to do
the check after does not seem worth it, when the work to be done
in the check is just incrementing a number, so simply remove the
check entirely and always update the offset even if unnecessary.