Separate extending hash cache from processing socket updates
getDifference reuses the same code to process socket updates.
This method used to also extend the hash cache with the new peers.

However, it turns out getDifference can return min peers,
which would cause the processing of updates to trigger a gap.
Because this gap would occur during getDifference, the code would
panic as intended.

This should help #135 once handling of update short is implemented.