Replace Notify with atomic counter
If two tasks race an invoke, both will believe the other did network.
This is why the notifier was introduced.

However, we can also solve this by getting a ticket number.

If, after acquiring the lock, it is still the task's turn,
it can drive the network.

If the turn changed while the lock was being acquired,
it must mean a different task drove the network.