Make Downloadable a trait
To accomodate for the changes, the implementation of DownloadIter
also has changed to use an enum instead, which better represents
all possible (exclusive) states.

It should be possible for Downloadable to guarantee a size, but
that is not a priority.

Redundant UserProfilePhoto type could also be deleted,
and ChatPhoto has been reworked to be a thin wrapper.