Propagate original update types further up
Instead of transforming AffectedMessages into UpdateDeleteMessages
directly in the sender, now multiple disjoint types can be returned
under the same enum that the session crate knows how to handle.

In particular, UpdateShortSentMessage can also be treated differently,
and this variant include the original request, enabling the session to
find out about the target peer and other information like the message
text, which it couldn't otherwise (which rendered these unusable).