Create SenderPool in mtsender
The Sender can be further refactored with its introduction.
Currently task spawning to bridge the old enqueuer is needed
but will become redundant. Dead connections do not cancel
pending requests either, and failing to connect panics.

There are some open questions such as reconnection policies
and how to best persist any of the new authentication keys.

Handling -404 errors and storing up-to-date DcOptions also
remains an open question (but this is not new).

For simplicity during the refactor, both proxy and WASM
support won't be added back until it's in a clean-enough state.