Update client to use new SenderPool
All connection stepping logic and downloader maps has been removed.
For now, authorization exports are also removed, but will need to be
brought back to support downloads and other features from other DCs.

Currently requests are cloned for every attempt, although this is not
too different from re-serializing them on failure. This should be
improved to either reuse the sent Vec or to Arc it.

Some logs have been worsened for now, but the constructor name can
still be inferred from the constructor ID in the future.

Update handling is left unanswered for the time being.

The naming on Client::connect no longer makes much sense since
connections are now on-demand, so it should probably be renamed.