Move all update state out of the client
Update state is unique, and there should only be one place handling it.
The client is intended to be easily shareable, so these two concepts do
not mix well.

Instead, a single update state instance can now be used with the
shareable client, which also gets rid of the need for locking around it.

For those who do not need updates, the stream can be dropped entirely
to avoid buffering them.