Delete ReconnectionPolicy from the sender
The new idea is that the SenderPool will do as ordered.
If the client attempts to invoke, the pool will take care of ensuring
there is an active connection to do so, and failures will be propagated.

All error retries should then be left for the client to decide what to
do with them, and whether to keep trying.

This may mean, once proxies are re-introduced, that it would be useful
to have a RetryPolicy on the Client capable of switching the proxy used
by the SenderPool. But this specific feature is not a priority.

The sender resetting logic may also be redundant now that fresh
connections are recreated trivially by the SenderPool.